refactor: 将 JavaScript 脚本整合为 index.mjs
更多信息:
fix?: 修复在Windows/Macos Runner上设置时区的问题
